fix: 관리자용 사용자 정보 상세 조회 API 수정

+++ src/main/java/com/flexrate/flexrate_back/member/domain/repository/MemberQueryRepository.java
@@ -15,5 +15,5 @@ public interface MemberQueryRepository {
 
     LoanApplication findLatestLoanApplication(Long memberId);
     Long countLoanTransactions(Long memberId);
-    Double findLatestInterestRate(Long memberId);
+    Float findLatestInterestRate(Long memberId);
 }
